Restructured reader class
The OOMMFData class was converted to a FieldData class and a new
MagnetisationData class was created inheriting the methods from the
FieldData class. MagnetisationData assumes omf files are the input and includes
the sk number calculation; the data from the m field is normalised in
this case. In the FieldData this normalisation is optional.